Remove invalid implementation of key enumeration
Changes:
 * Remove the KeyEnumerationMetric plugin
 * Remove the plugin registration
 * Remove the command in the CLI
 * Remove the no longer required TupleUInt from the argument list

Notes:

Key enumeration works by applying a CPA attack on the target then filter
the best key. The idea is to compute the pearson correlation matrix
then select the sample that is leaking the most. The old implementation
supposed the user would select range and then the algorithm would
perform a correlation between the corresponding range and the model
under key assumption